chore: remove unnecessary button component and add react-tabs
