+++ src/components/empty-active-recovery.tsx
@@ -1,4 +1,4 @@
-import Button from "./ui/button";
+import { Button } from "./ui/button";
 
 export default function EmptyActiveRecovery() {
   return (
