+++ src/components/ui/button.tsx
@@ -1,28 +1,85 @@
+import { cn } from "@/lib/utils";
+import { Slot } from "@radix-ui/react-slot";
+import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
-import { Button as BleuButton } from "./bleu-button";
-import { ButtonProps as BleuButtonProps } from "./bleu-button";
+import LoadingSpinner from "./loading-spinner";
 
-interface ButtonProps extends BleuButtonProps {
-  fullWidth?: boolean;
+const buttonVariants = cva(
+  "inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50",
+  {
+    variants: {
+      variant: {
+        default:
+          "bg-primary text-primary-foreground hover:bg-terciary hover:text-primary shadow",
+        destructive:
+          "bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90",
+        outline:
+          "border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground",
+        secondary:
+          "bg-secondary text-secondary-foreground shadow-sm hover:bg-secondary/80",
+        ghost: "hover:bg-accent hover:text-accent-foreground",
+        link: "text-primary underline-offset-4 hover:underline",
+      },
+      size: {
+        default: "h-9 px-4 py-2",
+        sm: "h-8 rounded-md px-3 text-xs",
+        lg: "h-10 rounded-md px-8",
+        icon: "h-9 w-9",
+      },
+    },
+    defaultVariants: {
+      variant: "default",
+      size: "default",
+    },
+  }
+);
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
+  loading?: boolean;
+  loadingText?: string;
 }
 
-export default function Button({
-  children,
-  className = "",
-  ...props
-}: ButtonProps) {
-  return (
-    <BleuButton
-      className={`
-        bg-primary 
-        text-primary-foreground 
-        hover:bg-terciary 
-        hover:text-primary 
-        ${className}
-      `}
-      {...props}
-    >
-      {children}
-    </BleuButton>
-  );
-}
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  (
+    {
+      className,
+      variant,
+      size,
+      asChild = false,
+      loading = false,
+      loadingText: _loadingText,
+      ...props
+    },
+    ref
+  ) => {
+    const Comp = asChild ? Slot : "button";
+    const loadingText = _loadingText ?? "Loading";
+
+    return (
+      <Comp
+        className={cn(buttonVariants({ variant, size }), className)}
+        ref={ref}
+        disabled={loading}
+        {...props}
+      >
+        {loading ? (
+          <>
+            <LoadingSpinner
+              aria-label={loadingText}
+              className="mr-2 size-4 animate-spin"
+            />
+            {loadingText}
+          </>
+        ) : (
+          props.children
+        )}
+      </Comp>
+    );
+  }
+);
+Button.displayName = "Button";
+
+export { Button, buttonVariants };
